Duplicating Cabin

+++ src/services/apiCabins.js
@@ -37,10 +37,13 @@ export async function createEditCabin(newCabin, id) {
     console.error(error);
     throw new Error('Cabins could not be created');
   }
+  // ---------  Upload image -----------
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
+
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
     console.error(storageError);
